refactor: ChessView 에서 Board 기능 분담
